Move plugins/ under the src/ directory.
