+++ index.js
@@ -2,7 +2,7 @@ import React from "react";
 import { render } from "react-dom";
 
 import KintoAdmin from "./src";
-import * as signoffPlugin from "./plugins/signoff";
+import * as signoffPlugin from "./src/plugins/signoff";
 
 
 const corePlugins = [
