+++ src/plugins/example/index.js
