+++ webpack.dev.js
@@ -39,7 +39,6 @@ module.exports = {
         include: [
           path.join(__dirname),
           path.join(__dirname, "src"),
-          path.join(__dirname, "plugins"),
         ],
       },
       {
