+++ webpack.prod.js
@@ -40,7 +40,6 @@ module.exports = {
         include: [
           path.join(__dirname),
           path.join(__dirname, "src"),
-          path.join(__dirname, "plugins"),
         ]
       },
       {
